게시판->게시글 이동 구현

+++ Final/src/main/java/com/test/project/domain/repository/BoardRepository.java
@@ -3,5 +3,5 @@ package com.test.project.domain.repository;
 import com.test.project.domain.entity.BoardEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<BoardEntity, Long> {
+public interface BoardRepository extends JpaRepository<BoardEntity, Integer> {
 }
